+++ mvp/src/main/java/by/vshkl/mvp/model/Comic.java
@@ -1,4 +1,4 @@
-package by.vshkl.model;
+package by.vshkl.mvp.model;
 
 public class Comic {
